refactor(agent): split task domain into lifecycle, storage sub-domains
