+++ packages/agent/src/task/index.ts
@@ -1,3 +1,5 @@
+// Task domain — lifecycle, storage, triggers
+
 import { Task } from "./types";
 export { Task };
 /** @deprecated Use Task.Run */
@@ -8,14 +10,8 @@ export type TaskRun = Task.Run;
 export const TriggerSignal = Task.TriggerSignal;
 /** @deprecated Use Task.TriggerSignal */
 export type TriggerSignal = Task.TriggerSignal;
+export * from "./lifecycle";
+export * from "./storage";
 export { TaskManager } from "./manager";
-export type { TriggerError, TriggerResult } from "./manager";
-export { TaskStorage, InMemoryTaskStore } from "./storage";
-export type { TaskStore } from "./storage";
-export { FileTaskStore } from "./file-task-storage";
-export { TaskStateMachine } from "./state-machine";
-export type { Checkpoint } from "./checkpoint";
-export { CheckpointManager } from "./checkpoint";
-export type { RecoveryAction, RecoveryResult } from "./recovery";
-export { CrashRecovery } from "./recovery";
+export type { TriggerError, TriggerResult } from "./trigger-engine";
 export { PolicyError } from "./errors";
