+++ packages/agent/src/task/lifecycle/index.ts
@@ -1,0 +1,9 @@
+export { TaskStatusManager, TaskStateMachine } from "./state-machine";
+
+export { CheckpointManager, type Checkpoint } from "./checkpoint";
+
+export {
+  CrashRecovery,
+  type RecoveryAction,
+  type RecoveryResult,
+} from "./recovery";
