+++ packages/agent/src/task/lifecycle/state-machine.ts
@@ -1,4 +1,4 @@
-import type { Task } from "./types";
+import type { Task } from "../types";
 
 /**
  * Task State Machine
